:sparkles: 4주차 웹소켓연결 및 ui구현완료
