+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/data/api/reseponse/MarketEvent.kt
@@ -1,0 +1,13 @@
+package io.seoj17.android_14th_compose_zzik.data.api.reseponse
+
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class MarketEvent(
+    @Json(name = "caution")
+    val caution: Caution,
+    @Json(name = "warning")
+    val warning: Boolean
+)
