+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/data/quilifier/NetworkQualifier.kt
@@ -1,0 +1,14 @@
+package io.seoj17.android_14th_compose_zzik.data.quilifier
+
+import javax.inject.Qualifier
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class UpBitBaseUrl
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class UpBitWebSocketUrl
+
+@Retention(AnnotationRetention.BINARY)
+@Qualifier
+annotation class UpBitWebSocketRequest
