+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/presentation/ui/CoinApp.kt
@@ -1,0 +1,7 @@
+package io.seoj17.android_14th_compose_zzik.presentation.ui
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class CoinApp : Application()
