+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/presentation/ui/component/CoinListButton.kt
@@ -1,0 +1,49 @@
+package io.seoj17.android_14th_compose_zzik.presentation.ui.component
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowLeft
+import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material3.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.semantics.Role
+
+@Composable
+fun CoinPreviousButton(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit = {},
+) {
+    Icon(
+        modifier = modifier.clickable(
+            role = Role.Button,
+            interactionSource = remember { MutableInteractionSource() },
+            indication = null,
+            onClick = onClick
+        ),
+        imageVector = Icons.Filled.KeyboardArrowLeft,
+        tint = Color.LightGray,
+        contentDescription = null,
+    )
+}
+
+@Composable
+fun CoinNextButton(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit = {},
+) {
+    Icon(
+        modifier = modifier.clickable(
+            role = Role.Button,
+            interactionSource = remember { MutableInteractionSource() },
+            indication = null,
+            onClick = onClick
+        ),
+        imageVector = Icons.Filled.KeyboardArrowRight,
+        tint = Color.LightGray,
+        contentDescription = null,
+    )
+}
