+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/presentation/ui/mypage/MyPageScreen.kt
@@ -1,0 +1,20 @@
+package io.seoj17.android_14th_compose_zzik.presentation.ui.mypage
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun MyPageScreen() {
+    Box(
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Text(
+            modifier = Modifier.align(alignment = Alignment.Center),
+            text = "내 정보 화면"
+        )
+    }
+}
