+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/presentation/ui/theme/Color.kt
@@ -1,0 +1,15 @@
+package io.seoj17.android_14th_compose_zzik.presentation.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+val CoinMainColor = Color(0xFF07275D)
+val CoinPriceUpColor = Color(0xFFC84A31)
+val CoinPriceDownColor = Color(0xFF1261C4)
