+++ app/src/main/java/io/seoj17/android_14th_compose_zzik/presentation/utils/ImmutableWrapper.kt
@@ -1,0 +1,6 @@
+package io.seoj17.android_14th_compose_zzik.presentation.utils
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+class ImmutableList<T>(val list: List<T> = emptyList())
