--- app/src/main/java/io/seoj17/android_14th_compose_zzik/ui/CoinApp.kt
@@ -1,7 +1,0 @@
-package io.seoj17.android_14th_compose_zzik.ui
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class CoinApp : Application()
