--- app/src/main/java/io/seoj17/android_14th_compose_zzik/ui/deposit/CoinDepositWithdrawal.kt
@@ -1,20 +1,0 @@
-package io.seoj17.android_14th_compose_zzik.ui.deposit
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-fun CoinDepositWithdrawal() {
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        Text(
-            modifier = Modifier.align(alignment = Alignment.Center),
-            text = "코인 입출금 화면"
-        )
-    }
-}
